front, test: hook useDurationInput
