+++ front/ui/ui-core/src/components/DurationInput/__tests__/useDurationInput.spec.tsx
@@ -1,197 +1,251 @@
-import { it, describe, expect } from 'vitest';
+import React, { useRef } from 'react';
 
+import { act, render, renderHook } from '@testing-library/react';
+import { it, describe, expect, vi, beforeEach } from 'vitest';
+
+import { toFormattedValues, type UnitProps, useDurationInput } from '../DurationInput';
 import {
-  normalizeUnits,
-  PRESETS,
-  toFormattedValues,
-  toTotalMilliseconds,
-  toUnitField,
-  type UnitProps,
-} from '..';
-
-const DEFAULT_PADCHAR = '0';
-const DEFAULT_MAX = Infinity;
-
-const DEFAULT_UNIT_KEYS = ['h', 'm', 's', 'ms'];
-const DEFAULT_UNIT_KEYS_COMBINATIONS: string[][] = [];
-
-for (let i = 0; i < DEFAULT_UNIT_KEYS.length; i++) {
-  for (let j = i + 2; j <= DEFAULT_UNIT_KEYS.length; j++) {
-    DEFAULT_UNIT_KEYS_COMBINATIONS.push(DEFAULT_UNIT_KEYS.slice(i, j));
-  }
-} // [['h', 'm'], ['m', 's'] ... ['h', 'm', 's']
-
-const DEFAULT_UNITS = normalizeUnits(
-  DEFAULT_UNIT_KEYS as UnitProps[],
+  DEFAULT_MAX,
   DEFAULT_PADCHAR,
-  DEFAULT_MAX
-);
+  DEFAULT_UNIT_KEYS,
+  DEFAULT_UNITS,
+  DEFAULT_VALUE,
+} from './consts';
 
-describe('utils', () => {
-  describe('normalizeUnits', () => {
-    it.each(DEFAULT_UNIT_KEYS)(
-      'should carry preset label and msPerUnit onto the output unit',
-      (key) => {
-        const result = normalizeUnits([key as UnitProps], DEFAULT_PADCHAR, DEFAULT_MAX);
-        expect(result[0].label).toBe(PRESETS[key].label);
-        expect(result[0].msPerUnit).toBe(PRESETS[key].msPerUnit);
-      }
+const mockOnChange = vi.fn();
+const mockBlur = vi.fn();
+const mockFocus = vi.fn();
+const mockPreventDefault = vi.fn();
+
+const useBuildDurationInputArgs = (
+  units: UnitProps[] = DEFAULT_UNIT_KEYS,
+  value = DEFAULT_VALUE,
+  padChar = DEFAULT_PADCHAR,
+  max = DEFAULT_MAX
+) => {
+  const containerRef = useRef<HTMLDivElement | null>(null);
+  return {
+    units,
+    value,
+    onChange: mockOnChange,
+    containerRef,
+    padChar,
+    max,
+  };
+};
+
+describe('useDurationInput', () => {
+  beforeEach(() => {
+    vi.clearAllMocks();
+  });
+
+  it('should initialize fields normalized', () => {
+    const { result } = renderHook(() => useDurationInput(useBuildDurationInputArgs()));
+    expect(result.current.fields).toStrictEqual(DEFAULT_UNITS);
+  });
+
+  it('should initialize formatted values', () => {
+    const { result } = renderHook(() => useDurationInput(useBuildDurationInputArgs()));
+    expect(result.current.formattedValues).toStrictEqual(
+      toFormattedValues(DEFAULT_UNITS, Infinity)
+    );
+  });
+
+  it.each(DEFAULT_UNITS)('should change formatted values according to changes', (unit) => {
+    const { result, rerender } = renderHook(() => useDurationInput(useBuildDurationInputArgs()));
+
+    result.current.handleChange(unit, '3');
+    rerender();
+    expect(result.current.formattedValues[unit.key]).toBe('3');
+    DEFAULT_UNITS.filter((u) => u.key !== unit.key).forEach((u) => {
+      expect(result.current.formattedValues[u.key]).toBe(
+        DEFAULT_PADCHAR.padStart(u.digits ?? 0, DEFAULT_PADCHAR)
+      );
+    });
+  });
+
+  it('should focus next field when active field is made full', async () => {
+    const { result } = renderHook(() => useDurationInput(useBuildDurationInputArgs()));
+
+    render(
+      result.current.fields.map((setting, idx) => (
+        <input
+          key={`${idx}`}
+          ref={(el) => {
+            if (el) result.current.inputRefs.current[setting.key] = el;
+          }}
+          onFocus={() => {
+            mockFocus(idx);
+          }}
+        />
+      ))
     );
 
-    it.each(DEFAULT_UNIT_KEYS_COMBINATIONS)(
-      'should set the max of a unit to the value of conversion',
-      (...units) => {
-        const result = normalizeUnits(units as UnitProps[], DEFAULT_PADCHAR, DEFAULT_MAX);
-        result.forEach((normalizedUnit, index) => {
-          const prev = index > 0 ? result[index - 1] : null;
-          if (prev) {
-            expect(normalizedUnit.max).toBe(Math.floor(prev.msPerUnit / normalizedUnit.msPerUnit));
-          } else {
-            expect(normalizedUnit.max).toBe(Math.floor(DEFAULT_MAX / normalizedUnit.msPerUnit));
-          }
-        });
-      }
+    await act(() => {
+      result.current.handleChange(
+        result.current.fields[1],
+        '3'.padStart(result.current.fields[1].digits!, '3')
+      );
+    });
+    expect(mockFocus).toHaveBeenCalledWith(2);
+  });
+
+  it('should not focus next field when last field is made full', async () => {
+    const { result } = renderHook(() => useDurationInput(useBuildDurationInputArgs()));
+
+    render(
+      result.current.fields.map((setting, idx) => (
+        <input
+          key={`${idx}`}
+          ref={(el) => {
+            if (el) result.current.inputRefs.current[setting.key] = el;
+          }}
+          onFocus={() => {
+            mockFocus(idx);
+          }}
+        />
+      ))
     );
-    it('should accept a custom UnitConfig object', () => {
-      const result = normalizeUnits(
-        [{ key: 'decaSeconds', label: 'ds', msPerUnit: 10_000, digits: 1 }],
-        '0',
-        10_000
-      );
-      expect(result[0].padChar).toBe(DEFAULT_PADCHAR);
-      expect(result[0].key).toBe('decaSeconds');
-      expect(result[0].msPerUnit).toBe(10_000);
-      expect(result[0].label).toBe('ds');
-      expect(result[0].digits).toBe(1);
-    });
 
-    it('should handle mixed string presets and custom objects', () => {
-      const result = normalizeUnits(
-        ['h', { key: 'frames', label: 'f', msPerUnit: 40, digits: 2 }],
-        '0',
-        DEFAULT_MAX
-      );
-      expect(result.map((u) => u.key)).toEqual(['h', 'frames']);
+    const lastField = result.current.fields.at(-1)!;
+    await act(() => {
+      result.current.handleChange(lastField, '3'.padStart(lastField.digits!, '3'));
     });
-    it('should always compute digits from max for the first unit, ignoring the custom digits parameter', () => {
-      const result = normalizeUnits(
-        [{ key: 'decaSeconds', label: 'ds', msPerUnit: 10_000, digits: 1 }],
-        '0',
-        10_000_000
-      );
-      expect(result[0].padChar).toBe(DEFAULT_PADCHAR);
-      expect(result[0].key).toBe('decaSeconds');
-      expect(result[0].msPerUnit).toBe(10_000);
-      expect(result[0].label).toBe('ds');
-      expect(result[0].digits).toBe(4);
-    });
-
-    it('should sort units from largest to smallest msPerUnit', () => {
-      const result = normalizeUnits(['ms', 'h', 's'], '0', DEFAULT_MAX);
-      expect(result.map((u) => u.key)).toEqual(['h', 's', 'ms']);
-    });
-
-    it('should propagate padChar to all units', () => {
-      const result = normalizeUnits(['h', 'm', 's'], '_', DEFAULT_MAX);
-      result.forEach((u) => expect(u.padChar).toBe('_'));
-    });
-
-    it('should accept only a label as a custom object unit', () => {
-      const result = normalizeUnits([{ label: 'h' }], DEFAULT_PADCHAR, DEFAULT_MAX);
-      const preset = PRESETS['h'];
-
-      expect(result[0].padChar).toBe(DEFAULT_PADCHAR);
-      expect(result[0].key).toBe('h');
-      expect(result[0].msPerUnit).toBe(preset.msPerUnit);
-      expect(result[0].label).toBe('h');
-      expect(result[0].digits).toBe(undefined);
-    });
-
-    it('should accept only a key as a custom object unit', () => {
-      const result = normalizeUnits([{ key: 'h' }], DEFAULT_PADCHAR, DEFAULT_MAX);
-      const preset = PRESETS['h'];
-
-      expect(result[0].padChar).toBe(DEFAULT_PADCHAR);
-      expect(result[0].key).toBe('h');
-      expect(result[0].msPerUnit).toBe(preset.msPerUnit);
-      expect(result[0].label).toBe('h');
-      expect(result[0].digits).toBe(undefined);
-    });
+    expect(mockFocus).not.toHaveBeenCalled();
   });
-  describe('toUnitField', () => {
-    it('for any number, should return th corresponding unit field', () => {
-      const result = toUnitField(10, {
-        label: 'ms',
-        key: 'ms',
-        msPerUnit: 1,
-        digits: 3,
-        max: 20,
-        padChar: '0',
+
+  it.each(['ArrowRight', ':', DEFAULT_UNIT_KEYS[0]])(
+    'should focus next field when pressing %s at last character of input',
+    async (key) => {
+      const { result } = renderHook(() => useDurationInput(useBuildDurationInputArgs()));
+
+      render(
+        result.current.fields.map((setting, idx) => (
+          <input
+            key={`${idx}`}
+            ref={(el) => {
+              if (el) result.current.inputRefs.current[setting.key] = el;
+            }}
+            onFocus={() => {
+              mockFocus(idx);
+            }}
+          />
+        ))
+      );
+
+      await act(() => {
+        result.current.handleKeyDown(result.current.fields[0], {
+          key,
+          preventDefault: () => {
+            mockPreventDefault();
+          },
+          currentTarget: {
+            selectionStart: 1,
+            value: 'a',
+          },
+        } as React.KeyboardEvent<HTMLInputElement>);
       });
-      expect(result).toStrictEqual(['ms', '010']);
+
+      expect(mockFocus).toHaveBeenCalledWith(1);
+      expect(mockPreventDefault).toHaveBeenCalled();
+    }
+  );
+
+  it('should focus previous field when pressing ArrowLeft at first character of input', async () => {
+    const { result } = renderHook(() => useDurationInput(useBuildDurationInputArgs()));
+
+    render(
+      result.current.fields.map((setting, idx) => (
+        <input
+          key={`${idx}`}
+          ref={(el) => {
+            if (el) result.current.inputRefs.current[setting.key] = el;
+          }}
+          onFocus={() => {
+            mockFocus(idx);
+          }}
+        />
+      ))
+    );
+
+    await act(() => {
+      result.current.handleKeyDown(result.current.fields[1], {
+        key: 'ArrowLeft',
+        preventDefault: () => {
+          mockPreventDefault();
+        },
+        currentTarget: {
+          selectionStart: 0,
+          value: 'a',
+        },
+      } as React.KeyboardEvent<HTMLInputElement>);
     });
 
-    it('for a number too high, should modulate the value at max', () => {
-      const result = toUnitField(10, {
-        label: 'ms',
-        key: 'ms',
-        msPerUnit: 1,
-        digits: 3,
-        max: 6,
-        padChar: '0',
-      });
-      expect(result).toStrictEqual(['ms', '004']);
-    });
+    expect(mockFocus).toHaveBeenCalledWith(0);
+    expect(mockPreventDefault).toHaveBeenCalled();
   });
 
-  describe('toFormattedValues', () => {
-    it.each([
-      {
-        ms: 3_600_000,
-        key: 'h',
-      },
-      {
-        ms: 60_000,
-        key: 'm',
-      },
-      {
-        ms: 1_000,
-        key: 's',
-      },
-      {
-        ms: 1,
-        key: 'ms',
-      },
-    ])('should format ($ms)ms as 1 ($key)', ({ ms, key }) => {
-      const result = toFormattedValues(DEFAULT_UNITS, ms);
-      const unit = DEFAULT_UNITS.find((u) => u.key === key);
+  it('should submit change when pressing enter', () => {
+    const { result, rerender } = renderHook(() => useDurationInput(useBuildDurationInputArgs()));
 
-      expect(unit).not.toBe(null);
+    result.current.handleChange(result.current.fields[0], '1');
 
-      expect(result[key]).toBe(String('1').padStart(unit?.digits ?? 0, unit?.padChar));
-    });
+    rerender();
+
+    result.current.handleKeyDown(result.current.fields[1], {
+      key: 'Enter',
+      currentTarget: {
+        selectionStart: 0,
+        value: '1',
+      },
+    } as React.KeyboardEvent<HTMLInputElement>);
+
+    expect(mockOnChange).toHaveBeenCalledWith(3_600_000);
   });
 
-  describe('toTotalMilliseconds', () => {
-    it.each([
-      {
-        values: { h: '1', m: '0', s: '0' },
-        total: 3_600_000,
-      },
-      {
-        values: { h: '1', m: '030', s: '0' },
-        total: 3_600_000 + 1_800_000,
-      },
-      {
-        values: { h: '0', m: '0', s: '10' },
-        total: 10_000,
-      },
-    ])('($values) should add up to ($total) milliseconds', ({ values, total }) => {
-      const result = toTotalMilliseconds(
-        DEFAULT_UNITS.filter((u) => Object.keys(values).includes(u.key)),
-        values
+  it('should handle focus', async () => {
+    const { result } = renderHook(() => useDurationInput(useBuildDurationInputArgs()));
+
+    const firstField = result.current.fields[0];
+
+    await act(() => {
+      render(
+        <input
+          key="firstInput"
+          ref={(el) => {
+            if (el) result.current.inputRefs.current[firstField.key] = el;
+          }}
+          onFocus={() => {
+            mockFocus();
+          }}
+        />
       );
-      expect(result).toBe(total);
+      result.current.handleFocus(firstField);
     });
+
+    expect(mockFocus).toHaveBeenCalled();
+  });
+
+  it('should handle blur', async () => {
+    const { result } = renderHook(() => useDurationInput(useBuildDurationInputArgs()));
+
+    const firstField = result.current.fields[0];
+
+    await act(() => {
+      render(
+        <input
+          key={`firstInput`}
+          ref={(el) => {
+            if (el) result.current.inputRefs.current[firstField.key] = el;
+          }}
+          onBlur={() => {
+            mockBlur();
+          }}
+        />
+      );
+    });
+
+    result.current.handleBlur({} as React.FocusEvent<HTMLInputElement>);
+    expect(mockOnChange).toHaveBeenCalled();
   });
 });
